requirements.txt
streamlit==1.41.1
anthropic==0.42.0
python-docx==1.2.0
pypdf==4.3.1
python-pptx==1.0.2
python-dotenv==1.2.2
openpyxl==3.1.5
SQLAlchemy==2.0.38
pgvector==0.3.6
psycopg[binary]==3.2.3
